+++ deepinv/physics/mri.py
@@ -172,7 +172,7 @@ class MRI(MRIMixin, DecomposablePhysics):
     The complex images :math:`x` and measurements :math:`y` should be of size (B, C,..., H, W) with C=2, where the first channel corresponds to the real part
     and the second channel corresponds to the imaginary part. The ``...`` is an optional depth dimension for 3D MRI data.
 
-    A fixed mask can be set at initialisation, or a new mask can be set either at forward (using ``physics(x, mask=mask)``) or using ``update_parameters``.
+    A fixed mask can be set at initialisation, or a new mask can be set either at forward (using ``physics(x, mask=mask)``) or using ``update``.
 
     .. note::
 
@@ -214,7 +214,7 @@ class MRI(MRIMixin, DecomposablePhysics):
         <BLANKLINE>
                  [[ 0.3744,  1.8622],
                   [ 0.0603, -0.6209]]]])
-        >>> physics.update_parameters(mask=mask) # Update mask on the fly
+        >>> physics.update(mask=mask) # Update mask on the fly
         >>> physics(x)
         tensor([[[[ 0.0000, -1.4290],
                   [ 0.4564, -0.0000]],
@@ -352,7 +352,7 @@ class MultiCoilMRI(MRIMixin, LinearPhysics):
         >>> physics(x).shape # B,C,N,H,W
         torch.Size([1, 2, 1, 2, 2])
         >>> coil_maps = torch.randn(1, 5, 2, 2, dtype=torch.complex64) # Define 5-coil sensitivity maps
-        >>> physics.update_parameters(coil_maps=coil_maps) # Update coil maps on the fly
+        >>> physics.update(coil_maps=coil_maps) # Update coil maps on the fly
         >>> physics(x).shape
         torch.Size([1, 2, 5, 2, 2])
 
@@ -516,7 +516,7 @@ class DynamicMRI(MRI, TimeMixin):
     The complex images :math:`x` and measurements :math:`y` should be of size (B, 2, T, H, W) where the first channel corresponds to the real part
     and the second channel corresponds to the imaginary part.
 
-    A fixed mask can be set at initialisation, or a new mask can be set either at forward (using ``physics(x, mask=mask)``) or using ``update_parameters``.
+    A fixed mask can be set at initialisation, or a new mask can be set either at forward (using ``physics(x, mask=mask)``) or using ``update``.
 
     .. note::
 
@@ -538,7 +538,7 @@ class DynamicMRI(MRI, TimeMixin):
         >>> x = torch.randn(1, 2, 2, 2, 2) # Define random video of shape (B,C,T,H,W)
         >>> mask = torch.rand_like(x) > 0.75 # Define random 4x subsampling mask
         >>> physics = DynamicMRI(mask=mask) # Physics with given mask
-        >>> physics.update_parameters(mask=mask) # Alternatively set mask on-the-fly
+        >>> physics.update(mask=mask) # Alternatively set mask on-the-fly
         >>> physics(x)
         tensor([[[[[-0.0000,  0.7969],
                    [-0.0000, -0.0000]],
@@ -646,7 +646,7 @@ class SequentialMRI(DynamicMRI):
     and thus have a fast pseudo-inverse and prox operators.
 
     A fixed mask can be set at initialisation, or a new mask can be set either at forward (using ``physics(x, mask=mask)``)
-    or using ``update_parameters``.
+    or using ``update``.
 
     .. note::
 
@@ -667,7 +667,7 @@ class SequentialMRI(DynamicMRI):
         >>> x = torch.randn(1, 2, 2, 2) # Define random image of shape (B,C,H,W)
         >>> mask = torch.zeros(1, 2, 3, 2, 2) # Empty demo time-varying mask with 3 frames
         >>> physics = SequentialMRI(mask=mask) # Physics with given mask
-        >>> physics.update_parameters(mask=mask) # Alternatively set mask on-the-fly
+        >>> physics.update(mask=mask) # Alternatively set mask on-the-fly
         >>> physics(x).shape # MRI sequential samples
         torch.Size([1, 2, 3, 2, 2])